Mandelbrot videogame
Now you can travel in the mandelbrot universe

Used SDL to open a window and handle user input

Fixed all requested changes in previous PR, aesthetic and logic ones. I also updated bmp.c
for the "write_bmp" function to be compatible. You can easily take a screenshot
of your favourite fractal spot.

Fixed even more requested changes. It's been a pleasure :) Gracias amigo!

+++ include/fractal.h
@@ -2,10 +2,10 @@
 #define FRACTAL_H
 
 struct world_coordinates {
-    double set_left;
-    double set_right;
-    double set_down;
-    double set_up;
+    long double set_left;
+    long double set_right;
+    long double set_down;
+    long double set_up;
 };
 
 void set_mandelbot(int world_x, int world_y, char *rgb, struct world_coordinates world);
